refactor(addressModal): fixed-flicked-image
